流程推进到create streamjob

+++ script/flink-prepare.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 hdfs dfs -rm -f "/streamingPro/$1"
 hdfs dfs -put "$1" /streamingPro
-rm $1 -f
+if [ $2 = "false" ]
+then
+    rm $1 -f
+fi
 hdfs dfs -ls /streamingPro
